@@ -4,8 +4,11 @@
     "keywords": [ "propel", "behavior", "event dispatcher", "event" ],
     "license": "MIT",
     "require": {
-        "symfony/event-dispatcher": "~2.1|~3.0|~3.4|^4.4|^5.4",
-        "propel/propel1": "^1.8"
+        "symfony/event-dispatcher": "^4.3|^5.0",
+        "dayspring-tech/propel1": "^1.8"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "~8.5.33|^9.0"
     },
     "authors": [
         {
@@ -15,5 +18,11 @@
     ],
     "autoload": {
         "classmap": ["src/"]
+    },
+    "scripts": {
+        "ci-test": [
+            "Composer\\Config::disableProcessTimeout",
+            "vendor/bin/phpunit -c ./phpunit.xml.dist"
+        ]
     }
 }
