@@ -3,4 +3,4 @@
 require_once __DIR__ . '/../vendor/autoload.php';
 set_include_path(__DIR__ . '/../vendor/phing/phing/classes' . PATH_SEPARATOR . get_include_path());
 
-require_once __DIR__ . '/../vendor/propel/propel1/generator/lib/util/PropelQuickBuilder.php';
+require_once __DIR__ . '/../vendor/dayspring-tech/propel1/generator/lib/util/PropelQuickBuilder.php';
